Move dialogs to widgets.
