+++ src/gui/context_menu.c
@@ -1,7 +1,7 @@
 #include "src/gui/actions.h"
 #include "src/gui/app_window.h"
-#include "src/gui/help_dialog.h"
-#include "src/gui/settings_dialog.h"
+#include "src/gui/widgets/help_dialog.h"
+#include "src/gui/widgets/settings_dialog.h"
 #include "src/lasr/auto-splitter.h"
 #include <gtk/gtk.h>
