+++ src/gui/widgets/help_dialog.h
