+++ src/gui/widgets/settings_dialog.h
